Overhaul logging
The logging used different approaches and by default didn't log much at all making it difficult to debug. This introduces a singleton that is used across all the applications and can be provided with either a normal logger or a syslog logger which should greatly simplify debugging.

Finally, the debug mode has been inverted to a quiet mode so that debug is logged by default (and can be switched off if desired).